requirements.txt
tweepy
feedparser
python-dotenv
PyYAML
beautifulsoup4
requests
Pillow
python-magic
